[structure] move video components to folder
