+++ src/components/VideoPlayer/IndicatorIcon.js
