+++ src/components/VideoPlayer/IndicatorIcon.module.scss
