+++ src/components/VideoPlayer/VideoContainer.module.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .fullscreenArea {
 	// height: 100%;
