+++ src/components/VideoPlayer/VideoContainerHeader.module.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .header {
 	display: flex;
