+++ src/components/VideoPlayer/VideoControl.js
