+++ src/components/VideoPlayer/VideoControl.module.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .control {
 	display: flex;
